+++ mithril/backends/with_autograd/jax_backend/utils.py
@@ -20,7 +20,7 @@ import numpy as np
 
 from .... import types
 from ....common import find_dominant_type
-from ....cores.jax.utils import dtype_map
+from ....cores.python.jax.utils import dtype_map
 from ...utils import DtypeSubTypes
 
 CODEGEN_CONFIG: dict[str, bool] = {
