+++ mithril/backends/with_autograd/mlx_backend/utils.py
@@ -22,7 +22,7 @@ import numpy as np
 
 from .... import types
 from ....common import find_dominant_type
-from ....cores.mlx.utils import dtype_map
+from ....cores.python.mlx.utils import dtype_map
 from ...utils import DtypeSubTypes
 
 CODEGEN_CONFIG: dict[str, bool] = {
