+++ mithril/cores/python/jax/ops.py
@@ -24,9 +24,10 @@ import jax.scipy.linalg as slin
 from jax import lax, vmap
 from jax import nn as functionals
 
-from ... import types
-from ...common import find_dominant_type
-from ..autograd_common_primitives import (
+from .... import types
+from ....common import find_dominant_type
+from ...utils import NestedFloatOrIntOrBoolList, is_tuple_int
+from ..common_primitives import (
     add,
     buffer,
     cartesian_diff,
@@ -69,7 +70,6 @@ from ..autograd_common_primitives import (
     tuple_converter,
     union,
 )
-from ..utils import NestedFloatOrIntOrBoolList, is_tuple_int
 from .utils import (
     broadcast_to_highest,
     calc_prob_matrix,
