+++ mithril/cores/python/mlx/utils.py
@@ -20,8 +20,8 @@ import mlx.core as mx
 import mlx.nn as nn
 import numpy as np
 
-from ...common import BiMap
-from ..utils import binary_search
+from ....common import BiMap
+from ...utils import binary_search
 
 dtype_map: BiMap[str, mx.Dtype] = BiMap(
     {
